requirements-web.txt
flet[web,cli]==0.27.6
flet-video==0.1.0
httpx>=0.28.1
screeninfo>=0.8.1
aiofiles>=24.1.0
streamget>=4.0.3
python-dotenv>=1.0.1
cachetools>=5.5.2